Integration into Skeleton
Got everything to compile. Will throw a lot of warnings, mostly about unused variables and probably other efficiency things in the all-inclusive files we're using.

Todo: Read RF, parse a message, send message based on some time interval (ensure the 'validity' of  the entire piece of data as being from the same 'packet' using getSafe? Or is there an existing system to prevent the message being sent in the middle of a 'write' operation?) , organize files with library?, kill a mockingbird (optional).